# frontend/openapi.json
{
 "components": {
  "schemas": {
...
   "AgentHorizonLevel": {
    "properties": {
     "horizon_h": {
      "title": "Horizon H",
      "type": "integer"
     },
     "level": {
      "$ref": "#/components/schemas/RiskLevelCode"
     },
     "probability": {
      "title": "Probability",
      "type": "number"
     }
    },
    "required": [
     "horizon_h",
     "level",
     "probability"
    ],
    "title": "AgentHorizonLevel",
    "type": "object"
   },
   "AgentProfile": {
    "properties": {
     "cash_capacity": {
      "title": "Cash Capacity",
      "type": "number"
     },
     "code": {
      "title": "Code",
      "type": "string"
     },
     "distributor_id": {
      "title": "Distributor Id",
      "type": "integer"
     },
     "district": {
      "title": "District",
      "type": "string"
     },
     "emoney_capacity": {
      "title": "Emoney Capacity",
      "type": "number"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "is_active": {
      "title": "Is Active",
      "type": "boolean"
     },
     "lat": {
      "title": "Lat",
      "type": "number"
     },
     "lng": {
      "title": "Lng",
      "type": "number"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "opened_on": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Opened On"
     },
     "region": {
      "title": "Region",
      "type": "string"
     },
     "tier": {
      "title": "Tier",
      "type": "integer"
     },
     "upazila": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Upazila"
     },
     "urban_rural": {
      "$ref": "#/components/schemas/UrbanRural"
     }
    },
    "required": [
     "id",
     "code",
     "name",
     "distributor_id",
     "region",
     "district",
     "upazila",
     "urban_rural",
     "tier",
     "lat",
     "lng",
     "cash_capacity",
     "emoney_capacity",
     "opened_on",
     "is_active"
    ],
    "title": "AgentProfile",
    "type": "object"
   },
...
   "AgentRisk": {
    "properties": {
     "agent_id": {
      "title": "Agent Id",
      "type": "integer"
     },
     "as_of": {
      "format": "date-time",
      "title": "As Of",
      "type": "string"
     },
     "by_horizon": {
      "items": {
       "$ref": "#/components/schemas/AgentHorizonLevel"
      },
      "title": "By Horizon",
      "type": "array"
     },
     "floats": {
      "items": {
       "$ref": "#/components/schemas/FloatRisk"
      },
      "title": "Floats",
      "type": "array"
     },
     "generated_at": {
      "format": "date-time",
      "title": "Generated At",
      "type": "string"
     },
     "level": {
      "$ref": "#/components/schemas/RiskLevelCode"
     },
     "model_version": {
      "title": "Model Version",
      "type": "string"
     }
    },
    "required": [
     "agent_id",
     "as_of",
     "model_version",
     "generated_at",
     "level",
     "by_horizon",
     "floats"
    ],
    "title": "AgentRisk",
    "type": "object"
   },
   "AgentRiskPage": {
    "properties": {
     "as_of": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "As Of"
     },
     "generated_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Generated At"
     },
     "horizon_h": {
      "title": "Horizon H",
      "type": "integer"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/AgentRiskRow"
      },
      "title": "Items",
      "type": "array"
     },
     "model_version": {
      "title": "Model Version",
      "type": "string"
     },
     "page": {
      "title": "Page",
      "type": "integer"
     },
     "page_size": {
      "title": "Page Size",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "page",
     "page_size",
     "horizon_h",
     "as_of",
     "model_version",
     "generated_at"
    ],
    "title": "AgentRiskPage",
    "type": "object"
   },
   "AgentRiskRow": {
    "properties": {
     "agent_id": {
      "title": "Agent Id",
      "type": "integer"
     },
     "code": {
      "title": "Code",
      "type": "string"
     },
     "district": {
      "title": "District",
      "type": "string"
     },
     "hours_to_stockout": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hours To Stockout"
     },
     "lat": {
      "title": "Lat",
      "type": "number"
     },
     "level": {
      "$ref": "#/components/schemas/RiskLevelCode"
     },
     "lng": {
      "title": "Lng",
      "type": "number"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "probability": {
      "title": "Probability",
      "type": "number"
     },
     "stockout_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Stockout At"
     },
     "tier": {
      "title": "Tier",
      "type": "integer"
     },
     "upazila": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Upazila"
     },
     "urban_rural": {
      "$ref": "#/components/schemas/UrbanRural"
     },
     "worst_float": {
      "$ref": "#/components/schemas/FloatType"
     }
    },
    "required": [
     "agent_id",
     "code",
     "name",
     "district",
     "upazila",
     "urban_rural",
     "tier",
     "lat",
     "lng",
     "level",
     "probability",
     "worst_float",
     "stockout_at",
     "hours_to_stockout"
    ],
    "title": "AgentRiskRow",
    "type": "object"
   },
   "AgentStockout": {
    "properties": {
     "agent_id": {
      "title": "Agent Id",
      "type": "integer"
     },
     "as_of": {
      "format": "date-time",
      "title": "As Of",
      "type": "string"
     },
     "floats": {
      "items": {
       "$ref": "#/components/schemas/FloatStockout"
      },
      "title": "Floats",
      "type": "array"
     },
     "generated_at": {
      "format": "date-time",
      "title": "Generated At",
      "type": "string"
     },
     "model_version": {
      "title": "Model Version",
      "type": "string"
     },
     "unit": {
      "const": "BDT",
      "default": "BDT",
      "title": "Unit",
      "type": "string"
     }
    },
    "required": [
     "agent_id",
     "as_of",
     "model_version",
     "generated_at",
     "floats"
    ],
    "title": "AgentStockout",
    "type": "object"
   },
   "AgentSummary": {
    "properties": {
     "agent": {
      "$ref": "#/components/schemas/AgentProfile"
     },
     "agent_id": {
      "title": "Agent Id",
      "type": "integer"
     },
     "as_of": {
      "format": "date-time",
      "title": "As Of",
      "type": "string"
     },
     "by_horizon": {
      "items": {
       "$ref": "#/components/schemas/AgentHorizonLevel"
      },
      "title": "By Horizon",
      "type": "array"
     },
     "floats": {
      "items": {
       "$ref": "#/components/schemas/FloatSummary"
      },
      "title": "Floats",
      "type": "array"
     },
     "generated_at": {
      "format": "date-time",
      "title": "Generated At",
      "type": "string"
     },
     "level": {
      "$ref": "#/components/schemas/RiskLevelCode"
     },
     "model_version": {
      "title": "Model Version",
      "type": "string"
     },
     "unit": {
      "const": "BDT",
      "default": "BDT",
      "title": "Unit",
      "type": "string"
     }
    },
    "required": [
     "agent_id",
     "as_of",
     "model_version",
     "generated_at",
     "agent",
     "level",
     "by_horizon",
     "floats"
    ],
    "title": "AgentSummary",
    "type": "object"
   },
...
   "FloatRisk": {
    "properties": {
     "float_type": {
      "$ref": "#/components/schemas/FloatType"
     },
     "horizons": {
      "items": {
       "$ref": "#/components/schemas/HorizonRisk"
      },
      "title": "Horizons",
      "type": "array"
     },
     "level": {
      "$ref": "#/components/schemas/RiskLevelCode"
     }
    },
    "required": [
     "float_type",
     "level",
     "horizons"
    ],
    "title": "FloatRisk",
    "type": "object"
   },
...
   "FloatType": {
    "enum": [
     "cash",
     "emoney"
    ],
    "title": "FloatType",
    "type": "string"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "RecommendationChannel": {
    "description": "How the money would reach the agent (app/rules/channel_rules.py).",
    "enum": [
     "swap",
     "top_up",
     "van",
     "self_fetch",
     "urgent_manual"
    ],
    "title": "RecommendationChannel",
    "type": "string"
   },
...
   "RequestAgent": {
    "properties": {
     "agent_id": {
      "title": "Agent Id",
      "type": "integer"
     },
     "code": {
      "title": "Code",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "agent_id",
     "code",
     "name"
    ],
    "title": "RequestAgent",
    "type": "object"
   },
   "RequestDecisionIn": {
    "properties": {
     "decision": {
      "enum": [
       "approve",
       "decline"
      ],
      "title": "Decision",
      "type": "string"
     },
     "note": {
      "maxLength": 500,
      "minLength": 1,
      "title": "Note",
      "type": "string"
     }
    },
    "required": [
     "decision",
     "note"
    ],
    "title": "RequestDecisionIn",
    "type": "object"
   },
   "RequestItem": {
    "properties": {
     "advisory": {
      "const": true,
      "default": true,
      "title": "Advisory",
      "type": "boolean"
     },
     "agent": {
      "$ref": "#/components/schemas/RequestAgent"
     },
     "amount_bdt": {
      "title": "Amount Bdt",
      "type": "number"
     },
     "channel": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RecommendationChannel"
       },
       {
        "type": "null"
       }
      ]
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "deadline_at": {
      "format": "date-time",
      "title": "Deadline At",
      "type": "string"
     },
     "decided_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Decided At"
     },
     "decided_by": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Decided By"
     },
     "float_type": {
      "$ref": "#/components/schemas/FloatType"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "recommendation_id": {
      "title": "Recommendation Id",
      "type": "integer"
     },
     "requested_by": {
      "format": "uuid",
      "title": "Requested By",
      "type": "string"
     },
     "status": {
      "$ref": "#/components/schemas/RequestStatus"
     }
    },
    "required": [
     "id",
     "recommendation_id",
     "agent",
     "float_type",
     "channel",
     "amount_bdt",
     "deadline_at",
     "status",
     "requested_by",
     "decided_by",
     "note",
     "created_at",
     "decided_at"
    ],
    "title": "RequestItem",
    "type": "object"
   },
   "RequestNoteIn": {
    "properties": {
     "note": {
      "anyOf": [
       {
        "maxLength": 500,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     }
    },
    "title": "RequestNoteIn",
    "type": "object"
   },
   "RequestPage": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/RequestItem"
      },
      "title": "Items",
      "type": "array"
     },
     "page": {
      "title": "Page",
      "type": "integer"
     },
     "page_size": {
      "title": "Page Size",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "RequestPage",
    "type": "object"
   },
   "RequestStatus": {
    "enum": [
     "requested",
     "approved",
     "declined",
     "fulfilled",
     "cancelled"
    ],
    "title": "RequestStatus",
    "type": "string"
   },
   "RiskLevelCode": {
    "enum": [
     "green",
     "amber",
     "red"
    ],
    "title": "RiskLevelCode",
    "type": "string"
   },
...
   "UrbanRural": {
    "enum": [
     "urban",
     "peri_urban",
     "rural"
    ],
    "title": "UrbanRural",
    "type": "string"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "title": "AgentPulse AI",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/agents": {
   "get": {
    "operationId": "list_agents_api_v1_agents_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/AgentProfile"
         },
         "title": "Response List Agents Api V1 Agents Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Agents",
    "tags": [
     "agents"
    ]
   }
  },
  "/api/v1/agents/risk": {
   "get": {
    "description": "Risk of the caller's agents at one horizon (agent: self; distributor: own agents).",
    "operationId": "list_risk_api_v1_agents_risk_get",
    "parameters": [
     {
      "description": "6, 24 or 72",
      "in": "query",
      "name": "horizon",
      "required": false,
      "schema": {
       "default": 24,
       "description": "6, 24 or 72",
       "title": "Horizon",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "level",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/RiskLevelCode"
        },
        {
         "type": "null"
        }
       ],
       "title": "Level"
      }
     },
     {
      "in": "query",
      "name": "sort",
      "required": false,
      "schema": {
       "default": "risk",
       "enum": [
        "risk",
        "stockout",
        "code",
        "name"
       ],
       "title": "Sort",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "maximum": 10000,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 20,
       "maximum": 100,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "q",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 80,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentRiskPage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Risk",
    "tags": [
     "risk"
    ]
   }
  },
  "/api/v1/agents/risk/export.csv": {
   "get": {
    "description": "The risk list (same filters, all pages) as CSV; scoped like GET /agents/risk.",
    "operationId": "export_risk_api_v1_agents_risk_export_csv_get",
    "parameters": [
     {
      "description": "6, 24 or 72",
      "in": "query",
      "name": "horizon",
      "required": false,
      "schema": {
       "default": 24,
       "description": "6, 24 or 72",
       "title": "Horizon",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "level",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/RiskLevelCode"
        },
        {
         "type": "null"
        }
       ],
       "title": "Level"
      }
     },
     {
      "in": "query",
      "name": "sort",
      "required": false,
      "schema": {
       "default": "risk",
       "enum": [
        "risk",
        "stockout",
        "code",
        "name"
       ],
       "title": "Sort",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "q",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 80,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "text/csv": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": "CSV file"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Export Risk",
    "tags": [
     "risk"
    ]
   }
  },
...
  "/api/v1/agents/{agent_id}/risk": {
   "get": {
    "description": "Stockout probability and green/amber/red level at 6 / 24 / 72 h per float.",
    "operationId": "get_risk_api_v1_agents__agent_id__risk_get",
    "parameters": [
     {
      "in": "path",
      "name": "agent_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Agent Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentRisk"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Risk",
    "tags": [
     "risk"
    ]
   }
  },
  "/api/v1/agents/{agent_id}/stockout": {
   "get": {
    "description": "Time-to-stockout + confidence per float (no refills assumed).",
    "operationId": "get_stockout_api_v1_agents__agent_id__stockout_get",
    "parameters": [
     {
      "in": "path",
      "name": "agent_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Agent Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentStockout"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Stockout",
    "tags": [
     "risk"
    ]
   }
  },
  "/api/v1/agents/{agent_id}/summary": {
   "get": {
    "description": "Profile, balances, time-to-stockout and risk levels in one call.",
    "operationId": "get_summary_api_v1_agents__agent_id__summary_get",
    "parameters": [
     {
      "in": "path",
      "name": "agent_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Agent Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentSummary"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Summary",
    "tags": [
     "risk"
    ]
   }
  },
...
  "/api/v1/recommendation-requests": {
   "get": {
    "description": "Requests in scope (agent: own; distributor: own agents; admin: all), newest first.",
    "operationId": "list_requests_api_v1_recommendation_requests_get",
    "parameters": [
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/RequestStatus"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "maximum": 10000,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 20,
       "maximum": 100,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestPage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Requests",
    "tags": [
     "recommendation-requests"
    ]
   }
  },
  "/api/v1/recommendation-requests/{request_id}/cancel": {
   "post": {
    "description": "Agent withdraws own request while it is still awaiting a decision (audit_log).",
    "operationId": "cancel_api_v1_recommendation_requests__request_id__cancel_post",
    "parameters": [
     {
      "in": "path",
      "name": "request_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Request Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/RequestNoteIn"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestItem"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Cancel",
    "tags": [
     "recommendation-requests"
    ]
   }
  },
  "/api/v1/recommendation-requests/{request_id}/decision": {
   "post": {
    "description": "Distributor approves or declines a requested item with a note (audit_log).",
    "operationId": "decide_api_v1_recommendation_requests__request_id__decision_post",
    "parameters": [
     {
      "in": "path",
      "name": "request_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Request Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RequestDecisionIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestItem"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Decide",
    "tags": [
     "recommendation-requests"
    ]
   }
  },
  "/api/v1/recommendation-requests/{request_id}/fulfil": {
   "post": {
    "description": "Distributor records that an approved request was delivered (audit_log).",
    "operationId": "fulfil_api_v1_recommendation_requests__request_id__fulfil_post",
    "parameters": [
     {
      "in": "path",
      "name": "request_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Request Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/RequestNoteIn"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestItem"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Fulfil",
    "tags": [
     "recommendation-requests"
    ]
   }
  },
  "/api/v1/recommendations/{recommendation_id}/request": {
   "post": {
    "description": "Agent asks the distributor to act on own recommendation. A repeat call returns the\nexisting request (200).",
    "operationId": "create_request_api_v1_recommendations__recommendation_id__request_post",
    "parameters": [
     {
      "in": "path",
      "name": "recommendation_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Recommendation Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestItem"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Create Request",
    "tags": [
     "recommendation-requests"
    ]
   }
  },
...
 }
}